requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
greenlet==2.0.2
jsonlines==3.1.0
playwright==1.36.0
pyee==9.0.4
python-dotenv==1.0.0
soupsieve==2.4.1
typing_extensions==4.7.1
